backend/requirements.txt
boto3==1.35.42
botocore==1.35.42
Flask==3.0.3
Flask_Cors==5.0.0
Requests==2.32.3
fake_headers>=1.0.2
pandas>=2.0.3
python-dotenv>=1.0.0
selenium>=4.12.0
webdriver_manager>=4.0.0

